+++ signdocs_brasil/__init__.py
@@ -60,7 +60,7 @@ from .response_metadata import ResponseMetadata
 from .token_cache import CachedToken, InMemoryTokenCache, TokenCache
 from .webhook_verifier import verify_webhook_signature
 
-__version__ = "1.3.0"
+__version__ = "1.4.1"
 
 __all__ = [
     # Client